+++ src/main/scala/com/blinkbox/books/spray/Directives.scala
@@ -15,6 +15,15 @@ case class Page(offset: Int, count: Int) {
   require(count > 0, "Count must be greater than 0")
 }
 
+object SortOrder {
+  val fieldParam = "order"
+  val descParam = "desc"
+}
+
+case class SortOrder(field: String, desc: Boolean) {
+  val asQueryParams = Seq((SortOrder.fieldParam, field.toString), (SortOrder.descParam, desc.toString))
+}
+
 trait Directives extends MonitoringDirectives {
 
   /**
@@ -27,22 +36,36 @@ trait Directives extends MonitoringDirectives {
 
   /**
    * Custom directive for extracting and validating page parameters (offset and count).
+   * @param defaultCount The default number of results per page.
    */
   def paged(defaultCount: Int) = parameters('offset.as[Int] ? 0, 'count.as[Int] ? defaultCount).as(Page)
+
+  /**
+   * Custom directive for extracting and validating sort order parameters (order and desc).
+   * @param defaultOrder The default sorting order.
+   */
+  def ordered(defaultOrder: SortOrder) = parameters(SortOrder.fieldParam ? defaultOrder.field, SortOrder.descParam.as[Boolean] ? defaultOrder.desc).as(SortOrder.apply _)
+
+  /**
+   * Custom directive for extracting and validating page and sort order parameters (offset, count, order, desc).
+   * @param defaultOrder The default sort order.
+   * @param defaultCount The default number of results per page.
+   */
+  def orderedAndPaged(defaultOrder: SortOrder, defaultCount: Int) = ordered(defaultOrder) & paged(defaultCount)
 
   /**
    * Directive for setting cache headers in a standard way.
    *
    * This sets time-dependent headers without requiring the use of a surrounding dynamic {} directive.
    *
-   * @param maxAge  The duration for which the returned value is cacheable.
+   * @param maxAge The duration for which the returned value is cacheable.
    */
   def cacheable(maxAge: Duration) = provideTimeNow.flatMap { now => withCacheHeaders(now, maxAge) }
 
   /**
    * Directive for completing request while also setting cache headers for the result.
    *
-   * @param maxAge  The duration for which the returned value is cacheable.
+   * @param maxAge The duration for which the returned value is cacheable.
    * @param response The response with which to complete the request.
    */
   def cacheable(maxAge: Duration, response: => ToResponseMarshallable): StandardRoute = cacheable(maxAge) & complete(response)
@@ -72,7 +95,6 @@ trait Directives extends MonitoringDirectives {
   private def withCacheHeaders(now: DateTime, maxAge: Duration) = respondWithHeaders(
     `Cache-Control`(`public`, `max-age`(maxAge.toSeconds)),
     RawHeader("Expires", (now + maxAge.toMillis).toRfc1123DateTimeString))
-
 }
 
 object Directives extends Directives