+++ src/main/scala/com/blinkbox/books/spray/v1/package.scala
@@ -1,6 +1,9 @@
 package com.blinkbox.books.spray
 
+import com.blinkbox.books.spray.Paging.PageLink
 import spray.http.{MediaType, MediaTypes}
+
+import scala.language.implicitConversions
 
 /**
  * Provides support for the version 1 media type, including common model objects.
@@ -16,4 +19,5 @@ package object v1 {
     binary = true, // binary as the encoding is defined as utf-8 by the json spec
     compressible = true))
 
+  implicit def pageLink2Link(pageLink: PageLink) = Link(pageLink.rel, pageLink.href, None, None)
 }